Update local storage for ArrowBuf

+++ tracdap-libs/tracdap-lib-data/src/main/java/org/finos/tracdap/common/storage/StorageErrors.java
@@ -191,7 +191,7 @@ public abstract class StorageErrors {
         }
     }
 
-    public ETrac chunkNotFullyWritten(int chunkBytes, int writtenBytes) {
+    public ETrac chunkNotFullyWritten(long chunkBytes, long writtenBytes) {
 
         try {
